Version 0.9S

+++ src/main/java/by/psu/gui/controllers/department_employee/all_item/ControllerItemEmployee.java
@@ -4,8 +4,12 @@ import by.psu.gui.LoaderFXML;
 import by.psu.gui.LoaderFilesFX;
 import by.psu.gui.logicalGui.ControllerFX;
 import by.psu.gui.logicalGui.ControllerFXLoader;
+import by.psu.logical.model.departure.Departure;
 import by.psu.logical.model.employee.Employee;
+import by.psu.logical.service.action.DepartureService;
 import by.psu.logical.service.employee_services.EmployeeService;
+import de.jensd.fx.glyphs.materialdesignicons.MaterialDesignIcon;
+import de.jensd.fx.glyphs.materialdesignicons.MaterialDesignIconView;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Label;
@@ -15,7 +19,10 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.paint.ImagePattern;
 import javafx.scene.shape.Circle;
 
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.net.URL;
+import java.util.List;
 import java.util.ResourceBundle;
 
 public class ControllerItemEmployee implements Initializable, ControllerFXLoader{
@@ -23,6 +30,7 @@ public class ControllerItemEmployee implements Initializable, ControllerFXLoader
     private EmployeeService es = new EmployeeService();
     private Employee employee = null;
     private ControllerItemPane parent = null;
+    private DepartureService departureService = new DepartureService();
 
     @FXML private Circle avatar;
     @FXML private AnchorPane anchorPane;
@@ -42,15 +50,34 @@ public class ControllerItemEmployee implements Initializable, ControllerFXLoader
 
     }
 
+    private boolean isAccessEmployee(){
+        List<Departure> departures = departureService.readALL();
+        for (Departure d : departures) {
+            if (!d.isDelete() && d.getStatus() != 3) {
+                if(d.getPostsEmployee().getEmployee().getId() == employee.getId())
+                    return false;
+            }
+        }
+        return true;
+    }
+
     @FXML
     private void actionDelete() {
-        es.delete(employee);
-        parent.deleteItem(anchorPane);
+        if(isAccessEmployee()) {
+            es.deleteEmployee(employee.getId());
+            parent.deleteItem(anchorPane);
+        } else {
+            ControllerItemPane.getControllerItemPane().message("Действие заморожено. Сотрудник имеет выезды.");
+        }
     }
 
     @FXML
     private void actionEdit() {
-        LoaderFXML.loaderController("/gui.resources/personal_department/item_employee/stack_pane_employee_v2.fxml", parent.getControllerDeparOder().getWorkspace(), parent.getControllerDeparOder(), employee);
+        if(isAccessEmployee()) {
+            LoaderFXML.loaderController("/gui.resources/personal_department/item_employee/stack_pane_employee_v2.fxml", parent.getControllerDeparOder().getWorkspace(), parent.getControllerDeparOder(), employee);
+        } else {
+            ControllerItemPane.getControllerItemPane().message("Действие заморожено. Сотрудник имеет выезды.");
+        }
     }
 
     @FXML
@@ -71,6 +98,12 @@ public class ControllerItemEmployee implements Initializable, ControllerFXLoader
                         + employee.getLastName() + " "
                         + employee.getPatronymic()
         );
+        if(employee.getAvatar()!=null){
+            try {
+                Image image = new Image(new FileInputStream(employee.getAvatar()));
+                avatar.setFill(new ImagePattern(image));
+            } catch (FileNotFoundException ignored) {}
+        }
     }
 
     @Override
